Make sysytem start configurable are verbose.
Since xv6 is a learning OS it should be a bit more verbose in the booting process.
The Count of the CPU's and the RAM can now be configured at the HEAD of  Makefile and in kernel/param.h the MAXMEM can be defined as well <- Should be the less equal to the Number in the Makefile.
It is prepared to change the CPUSTACKSIZE in the future to play a bit with the internals of the implemtation